+++ packages/expo-yarn-workspaces/index.js
@@ -2,18 +2,12 @@
 
 const { getDefaultConfig } = require('@expo/metro-config');
 const debug = require('debug')('workspaces');
-const findYarnWorkspaceRoot = require('find-yarn-workspace-root');
 const path = require('path');
-
-const getSymlinkedNodeModulesForDirectory = require('./common/get-symlinked-modules');
 
 /**
  * Returns a configuration object in the format expected for "metro.config.js" files. The
  * configuration:
  *
- *   * includes the Yarn workspace root in Metro's list of root directories
- *   * resolves symlinked packages, namely workspaces
- *   * excludes all modules from Haste's module system (providesModule)
  *   * excludes modules in the native Android and Xcode projects
  */
 exports.createMetroConfiguration = function createMetroConfiguration(projectPath, options) {
@@ -25,43 +19,15 @@ exports.createMetroConfiguration = function createMetroConfiguration(projectPath
     ...defaultConfig
   } = getDefaultConfig(projectPath, options);
 
-  let watchFolders;
-  let extraNodeModules;
-
-  const workspaceRootPath = findYarnWorkspaceRoot(projectPath);
-  if (workspaceRootPath) {
-    debug(`Found Yarn workspace root at %s`, workspaceRootPath);
-    watchFolders = [workspaceRootPath];
-    extraNodeModules = {
-      ...getSymlinkedNodeModulesForDirectory(workspaceRootPath),
-      ...getSymlinkedNodeModulesForDirectory(projectPath),
-    };
-  } else {
-    debug(`Could not find Yarn workspace root`);
-    watchFolders = [];
-    extraNodeModules = getSymlinkedNodeModulesForDirectory(projectPath);
-  }
-
   return {
     ...defaultConfig,
     // Search for modules from the project's root directory
     projectRoot: projectPath,
 
-    // Include npm packages from the workspace root, where packages are hoisted
-    watchFolders,
     resolver: {
       ...defaultConfig.resolver,
       // test-suite includes a db asset
       assetExts: [...defaultConfig.resolver.assetExts, 'db'],
-
-      // Include .cjs files
-      sourceExts: [...defaultConfig.resolver.sourceExts, 'cjs'],
-
-      // Make the symlinked packages visible to Metro
-      extraNodeModules,
-
-      // Use Node-style module resolution instead of Haste everywhere
-      providesModuleNodeModules: [],
 
       // Ignore test files and JS files in the native Android and Xcode projects
       blockList: [
@@ -74,6 +40,10 @@ exports.createMetroConfiguration = function createMetroConfiguration(projectPath
     transformer: {
       ...defaultConfig.transformer,
       // Ignore file-relative Babel configurations and apply only the project's
+      // NOTE: The Metro transformer still searches for and uses .babelrc and .babelrc.js files:
+      // https://github.com/facebook/react-native/blob/753bb2094d95c8eb2152d2a2c1f0b67bbeec36de/packages/react-native-babel-transformer/src/index.js#L81
+      // This is in contrast with Babel, which reads only babel.config.json before evaluating its
+      // "babelrc" option: https://babeljs.io/docs/options#configfile
       enableBabelRCLookup: false,
     },
   };